Rename test files to include 'tests' directory in the path
